+++ meerschaum/utils/sql.py
@@ -376,7 +376,7 @@ columns_indices_queries = {
         WHERE
             t.name IN ('{table}', '{table_trunc}')
             AND s.name = '{schema}'
-            AND i.type IN (1, 2)  -- 1 = CLUSTERED, 2 = NONCLUSTERED
+            AND i.type IN (1, 2)
     """,
     'oracle': """
         SELECT
@@ -590,7 +590,14 @@ def dateadd_str(
     from meerschaum.utils.dtypes.sql import get_db_type_from_pd_type, get_pd_type_from_db_type
     dateutil_parser = attempt_import('dateutil.parser')
     if 'int' in str(type(begin)).lower():
-        return str(begin)
+        num_str = str(begin)
+        if number is not None and number != 0:
+            num_str += (
+                f' + {number}'
+                if number > 0
+                else f" - {number * -1}"
+            )
+        return num_str
     if not begin:
         return ''
 
@@ -797,7 +804,7 @@ def sql_item_name(item: str, flavor: str, schema: Optional[str] = None) -> str:
     """
     truncated_item = truncate_item_name(str(item), flavor)
     if flavor == 'oracle':
-        truncated_item = pg_capital(truncated_item)
+        truncated_item = pg_capital(truncated_item, quote_capitals=True)
         ### NOTE: System-reserved words must be quoted.
         if truncated_item.lower() in (
             'float', 'varchar', 'nvarchar', 'clob',
@@ -824,14 +831,17 @@ def sql_item_name(item: str, flavor: str, schema: Optional[str] = None) -> str:
     return schema_prefix + wrappers[0] + truncated_item + wrappers[1]
 
 
-def pg_capital(s: str) -> str:
+def pg_capital(s: str, quote_capitals: bool = True) -> str:
     """
     If string contains a capital letter, wrap it in double quotes.
     
     Parameters
     ----------
-    s: str :
-    The string to be escaped.
+    s: str
+        The string to be escaped.
+
+    quote_capitals: bool, default True
+        If `False`, do not quote strings with contain only a mix of capital and lower-case letters.
 
     Returns
     -------
@@ -845,16 +855,24 @@ def pg_capital(s: str) -> str:
     'my_table'
 
     """
-    if '"' in s:
+    if s.startswith('"') and s.endswith('"'):
         return s
+
+    s = s.replace('"', '')
+
     needs_quotes = s.startswith('_')
-    for c in str(s):
-        if ord(c) < ord('a') or ord(c) > ord('z'):
-            if not c.isdigit() and c != '_':
+    if not needs_quotes:
+        for c in s:
+            if c == '_':
+                continue
+
+            if not c.isalnum() or (quote_capitals and c.isupper()):
                 needs_quotes = True
                 break
+
     if needs_quotes:
         return '"' + s + '"'
+
     return s
 
 
@@ -1448,6 +1466,7 @@ def get_update_queries(
     schema: Optional[str] = None,
     patch_schema: Optional[str] = None,
     identity_insert: bool = False,
+    null_indices: bool = True,
     debug: bool = False,
 ) -> List[str]:
     """
@@ -1488,6 +1507,9 @@ def get_update_queries(
     identity_insert: bool, default False
         If `True`, include `SET IDENTITY_INSERT` queries before and after the update queries.
         Only applies for MSSQL upserts.
+
+    null_indices: bool, default True
+        If `False`, do not coalesce index columns before joining.
 
     debug: bool, default False
         Verbosity toggle.
@@ -1579,11 +1601,17 @@ def get_update_queries(
 
     coalesce_join_cols_str = ', '.join(
         [
-            'COALESCE('
-            + sql_item_name(c_name, flavor)
-            + ', '
-            + get_null_replacement(c_type, flavor)
-            + ')'
+            (
+                (
+                    'COALESCE('
+                    + sql_item_name(c_name, flavor)
+                    + ', '
+                    + get_null_replacement(c_type, flavor)
+                    + ')'
+                )
+                if null_indices
+                else sql_item_name(c_name, flavor)
+            )
             for c_name, c_type in join_cols_types
         ]
     )
@@ -1633,19 +1661,29 @@ def get_update_queries(
     def and_subquery(l_prefix: str, r_prefix: str):
         return '\n            AND\n                '.join([
             (
-                "COALESCE("
-                + l_prefix
-                + sql_item_name(c_name, flavor, None)
-                + ", "
-                + get_null_replacement(c_type, flavor)
-                + ")"
-                + '\n                =\n                '
-                + "COALESCE("
-                + r_prefix
-                + sql_item_name(c_name, flavor, None)
-                + ", "
-                + get_null_replacement(c_type, flavor)
-                + ")"
+                (
+                    "COALESCE("
+                    + l_prefix
+                    + sql_item_name(c_name, flavor, None)
+                    + ", "
+                    + get_null_replacement(c_type, flavor)
+                    + ")"
+                    + '\n                =\n                '
+                    + "COALESCE("
+                    + r_prefix
+                    + sql_item_name(c_name, flavor, None)
+                    + ", "
+                    + get_null_replacement(c_type, flavor)
+                    + ")"
+                )
+                if null_indices
+                else (
+                    l_prefix
+                    + sql_item_name(c_name, flavor, None)
+                    + ' = '
+                    + r_prefix
+                    + sql_item_name(c_name, flavor, None)
+                )
             ) for c_name, c_type in join_cols_types
         ])